requirements.txt
aniso8601==9.0.1
attrs==25.3.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
cryptography==44.0.0
decorator==5.1.1
Flask==3.1.0
Flask-HTTPAuth==4.8.0
Flask-RESTful==0.3.10
gevent==24.11.1
greenlet==3.1.1
grequests==0.7.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpath-rw==1.4.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multipart==1.2.1
ocptv==0.1.7
packaging==24.2
paramiko==3.5.0
pluggy==1.5.0
ply==3.11
prettytable==3.16.0
pycparser==2.22
Pygments==2.19.1
PyNaCl==1.5.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-sugar==1.0.0
pytz==2024.2
PyYAML==6.0.2
redfish==3.2.6
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
requests-unixsocket==0.3.0
responses==0.25.3
rich==13.9.4
rpds-py==0.24.0
setuptools==75.6.0
six==1.17.0
sshtunnel==0.4.0
StringGenerator==0.4.4
tabulate==0.9.0
termcolor==2.5.0
typing_extensions==4.13.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
